# contracts/openapi/core.v1.yaml
openapi: 3.1.0
info:
  title: ESWMP Core API
  version: v1
  description: |
    Scheduling primitives: Resources, Calendars, Availability Rules, Reservations,
    and Appointments. This service owns scheduling and nothing else — it never
    stores or returns any data belonging to the caller's own domain beyond the
    opaque externalReferenceType/externalReferenceId pair. See docs/ESWMP_VISION.md
    "Key rule".
servers:
  - url: http://localhost:6001
    description: Local development
security:
  - BearerAuth: []
paths:
  /api/v1/resources:
    post:
      tags: [Resources]
      summary: Register a Resource (a person, vehicle, room, or piece of equipment that can be scheduled)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateResourceRequest'
      responses:
        '201':
          description: Resource created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
    get:
      tags: [Resources]
      summary: List resources for the tenant
      parameters:
        - name: resourceType
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/ResourceStatus' }
      responses:
        '200':
          description: Paged resource list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedResourceResult'
  /api/v1/resources/{id}:
    get:
      tags: [Resources]
      summary: Get a single Resource
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Resource detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
  /api/v1/resources/{id}/activate:
    post:
      tags: [Resources]
      summary: Activate a Resource (Draft/PendingVerification/Inactive -> Active)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResourceLifecycleRequest' }
      responses:
        '200': { description: Resource activated, content: { application/json: { schema: { $ref: '#/components/schemas/Resource' } } } }
        '409': { description: Invalid transition from current status }
        '412': { description: expectedVersion did not match current Version }
  /api/v1/resources/{id}/suspend:
    post:
      tags: [Resources]
      summary: Suspend an Active Resource
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResourceLifecycleRequest' }
      responses:
        '200': { description: Resource suspended, content: { application/json: { schema: { $ref: '#/components/schemas/Resource' } } } }
        '409': { description: Invalid transition from current status }
        '412': { description: expectedVersion did not match current Version }
  /api/v1/resources/{id}/reactivate:
    post:
      tags: [Resources]
      summary: Reactivate a Suspended Resource
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResourceLifecycleRequest' }
      responses:
        '200': { description: Resource reactivated, content: { application/json: { schema: { $ref: '#/components/schemas/Resource' } } } }
        '409': { description: Invalid transition from current status }
        '412': { description: expectedVersion did not match current Version }
  /api/v1/resources/{id}/deactivate:
    post:
      tags: [Resources]
      summary: Deactivate an Active or Suspended Resource
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResourceLifecycleRequest' }
      responses:
        '200': { description: Resource deactivated, content: { application/json: { schema: { $ref: '#/components/schemas/Resource' } } } }
        '409': { description: Invalid transition from current status }
        '412': { description: expectedVersion did not match current Version }
  /api/v1/resources/{id}/retire:
    post:
      tags: [Resources]
      summary: Retire a Resource permanently (terminal state)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResourceLifecycleRequest' }
      responses:
        '200': { description: Resource retired, content: { application/json: { schema: { $ref: '#/components/schemas/Resource' } } } }
        '409': { description: Invalid transition from current status }
        '412': { description: expectedVersion did not match current Version }
  /api/v1/resources/{id}/capabilities:
    post:
      tags: [Resources]
      summary: Grant a capability to a Resource
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddResourceCapabilityRequest' }
      responses:
        '201': { description: Capability added, content: { application/json: { schema: { $ref: '#/components/schemas/ResourceCapability' } } } }
  /api/v1/resources/{id}/skills:
    post:
      tags: [Resources]
      summary: Add a skill to a Resource
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddResourceSkillRequest' }
      responses:
        '201': { description: Skill added, content: { application/json: { schema: { $ref: '#/components/schemas/ResourceSkill' } } } }
  /api/v1/resources/{id}/certifications:
    post:
      tags: [Resources]
      summary: Add a certification to a Resource
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddResourceCertificationRequest' }
      responses:
        '201': { description: Certification added, content: { application/json: { schema: { $ref: '#/components/schemas/ResourceCertification' } } } }
...
  /api/v1/reservations:
    post:
      tags: [Reservations]
      summary: Create a temporary hold on a Resource for a time window
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReservationRequest'
      responses:
        '201':
          description: Reservation held
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        '409':
          description: The Resource already has an overlapping Held or Confirmed reservation for the requested time window
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
  /api/v1/reservations/{id}/confirm:
    post:
      tags: [Reservations]
      summary: Confirm a held reservation (creates the Appointment)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Reservation confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
  /api/v1/reservations/{id}/cancel:
    post:
      tags: [Reservations]
      summary: Cancel a held or confirmed reservation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Reservation cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ResourceStatus:
      type: string
      description: >-
        Expanded from the original 3-state (Active/Inactive/Suspended) enum to a
        full onboarding-through-retirement lifecycle.
      enum: [Draft, PendingVerification, Active, Suspended, Inactive, Retired]
    VerificationStatus:
      type: string
      enum: [NotRequired, NotStarted, Pending, Verified, Rejected, Expired]
    CreateResourceRequest:
      type: object
      required: [resourceType, name, timezone]
      properties:
        resourceType:
          type: string
          description: Caller-defined type label, e.g. "GroomingVan", "MeetingRoom", "DeliveryDriver"
        name: { type: string }
        timezone: { type: string, example: "America/Toronto" }
        capacity: { type: integer, nullable: true }
        skills:
          type: array
          items: { type: string }
        locationLatitude: { type: number, format: double, nullable: true }
        locationLongitude: { type: number, format: double, nullable: true }
    Resource:
      type: object
      required: [id, tenantId, resourceType, name, timezone, status, verificationStatus, version]
      properties:
        id: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        resourceType: { type: string }
        resourceTypeId: { type: string, format: uuid, nullable: true, description: "Optional link to a tenant-defined ResourceType row." }
        name: { type: string }
        timezone: { type: string }
        status: { $ref: '#/components/schemas/ResourceStatus' }
        verificationStatus: { $ref: '#/components/schemas/VerificationStatus' }
        version: { type: integer, description: "Optimistic-concurrency token for lifecycle transitions." }
        capacity: { type: integer, nullable: true }
        skills:
          type: array
          items: { type: string }
        location:
          $ref: '#/components/schemas/GeoPoint'
    ResourceLifecycleRequest:
      type: object
      properties:
        reasonCode: { type: string, nullable: true }
        comment: { type: string, nullable: true }
        expectedVersion: { type: integer, nullable: true, description: "If supplied and it doesn't match the Resource's current version, returns 412." }
    AddResourceCapabilityRequest:
      type: object
      required: [capabilityCode, level]
      properties:
        capabilityCode: { type: string, example: "HEAVY_LIFTING" }
        level: { type: integer, minimum: 1, maximum: 4, default: 1 }
        effectiveFrom: { type: string, format: date, nullable: true }
        effectiveTo: { type: string, format: date, nullable: true }
    ResourceCapability:
      type: object
      required: [id, resourceId, capabilityCode, level, status]
      properties:
        id: { type: string, format: uuid }
        resourceId: { type: string, format: uuid }
        capabilityCode: { type: string }
        level: { type: integer }
        status: { type: string, enum: [Pending, Active, Suspended, Expired, Revoked] }
        effectiveFrom: { type: string, format: date, nullable: true }
        effectiveTo: { type: string, format: date, nullable: true }
    AddResourceSkillRequest:
      type: object
      required: [skillCode, level]
      properties:
        skillCode: { type: string }
        level: { type: integer, minimum: 1, default: 1 }
        yearsOfExperience: { type: number, nullable: true }
        verified: { type: boolean, default: false }
    ResourceSkill:
      type: object
      required: [id, resourceId, skillCode, level, verified]
      properties:
        id: { type: string, format: uuid }
        resourceId: { type: string, format: uuid }
        skillCode: { type: string }
        level: { type: integer }
        yearsOfExperience: { type: number, nullable: true }
        verified: { type: boolean }
    AddResourceCertificationRequest:
      type: object
      required: [certificationTypeCode]
      properties:
        certificationTypeCode: { type: string }
        credentialReference: { type: string, nullable: true }
        issuedAt: { type: string, format: date, nullable: true }
        expiresAt: { type: string, format: date, nullable: true }
    ResourceCertification:
      type: object
      required: [id, resourceId, certificationTypeCode, status]
      properties:
        id: { type: string, format: uuid }
        resourceId: { type: string, format: uuid }
        certificationTypeCode: { type: string }
        credentialReference: { type: string, nullable: true }
        issuedAt: { type: string, format: date, nullable: true }
        expiresAt: { type: string, format: date, nullable: true }
        status: { type: string, enum: [Submitted, PendingVerification, Valid, Rejected, Expired, Revoked] }
    GeoPoint:
      type: object
      nullable: true
      properties:
        latitude: { type: number, format: double }
        longitude: { type: number, format: double }
    PagedResourceResult:
      type: object
      required: [items, page, pageSize, totalCount]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Resource' }
        page: { type: integer }
        pageSize: { type: integer }
        totalCount: { type: integer }
...
    CreateReservationRequest:
      type: object
      required: [resourceId, startTime, endTime, externalReferenceType, externalReferenceId]
      properties:
        resourceId: { type: string, format: uuid }
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time }
        holdDurationMinutes: { type: integer, default: 15 }
        externalReferenceType:
          type: string
          description: Opaque caller-domain type label, e.g. "booking". ESWMP never inspects this.
        externalReferenceId:
          type: string
          description: Opaque caller-domain identifier, e.g. "BK-12345". ESWMP never inspects this.
    Reservation:
      type: object
      required: [id, tenantId, resourceId, startTime, endTime, status, externalReferenceType, externalReferenceId]
      properties:
        id: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        resourceId: { type: string, format: uuid }
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time }
        status: { type: string, enum: [Held, Confirmed, Expired, Cancelled] }
        expiresAt: { type: string, format: date-time, nullable: true }
        externalReferenceType: { type: string }
        externalReferenceId: { type: string }
...